requirements.txt
requests>=2.32
python-dotenv>=1.0
sentence-transformers>=2.7
faiss-cpu>=1.7.4          # Windows build
tenacity>=8.2
openai>=1.27
pyreadline3
PyJWT>=2.8
msal>=1.28

# torch / torchvision installed separately per-GPU instructions 
# CUDA 12.1 – adjust index URL if your driver is older (cu118, cu116, …)
# pip install torch torchvision --extra-index-url https://download.pytorch.org/whl/cu121
